+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is used to list changes made in each version of the firewall cookbook.
 
 - Support for firewalld 2.0.0 and the platforms that use it; RHEL 10 and Ubuntu 24.04.
   - `priority`, `ingress_priority`, `egress_priority` properties added to `firewalld_zone`.
+- Added `firewalld_rich_rule` resource for adding/removing rich rules to/from firewalld zones.
 
 ### Changed